Add Robot that can rotate and move forward
The robot starts by facing South, this has been arbitrarily decided,
could as well be any other direction.

For now Robot, Direction and Position are encapsulated within the Robot
module, could be considered for extraction later if it made sense.

This commit also adds a few unit tests sanity checking the logic. Not
the most straightforward or readable way to test, but it gets the job
done without exposing unnecessary functions from the Robot module.