Fix robot rotation animation spinning too much
Before, because the animations are happening with transition, when the
rotation would change from 0 to 335 or viceversa the robot would do
a fast spin instead of rotating the shortest distance.

To fix it we store the rotation degrees in the robot and update them
appropriately internally when the robot is rotated. This way we keep the
rotation close to the previous one when re-rendering and the robot
doesn't over-spin when animating.